requirements.txt
streamlit
yfinance
pandas
requests
beautifulsoup4
matplotli
flask
